% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/great_circle_distance.R
\name{great_circle_distance}
\alias{great_circle_distance}
\title{great_circle_distance}
\usage{
great_circle_distance(lat1, lat2, lon1, lon2, scale = 6371)
}
\arguments{
\item{lat1}{Latitude of Position 1}

\item{lat2}{Latitude of Position 2}

\item{lon1}{Longitude of Position 1}

\item{lon2}{Longitude of Position 2}

\item{scale}{Scale of the Planet}
}
\value{
Great circle distance.
}
\description{
great_circle_distance
}
